Replacement of the MM heap process.
Old way sucks and wasn't going to work.

New way is a a flat list of allocations, allocations are linked together to keep the order of addresses sorted. Each time we need a new allocation structure, we need to scan the list until we find an unallocated entry. Then it gets linked to the correct position in the linked list.  Added some blurb to explain in the heap_memory.c file.

Work in progress.

+++ types.h
@@ -4,7 +4,7 @@
 #define TRUE 1
 #define FALSE 0
 
-typedef int BOOL;
+typedef short BOOL;
 
 
 #endif
